STEP-4 implement create order

+++ src/main/java/com/instructure/shop/promotion/PromotionAggregator.java
@@ -28,7 +28,7 @@ public class PromotionAggregator {
    * @return the cheapest cost for provided courses
    */
   public BigDecimal getTheLowestCost(
-      Map<Course, Long> quantityByCourse,
+      Map<Course, Integer> quantityByCourse,
       List<Promotion> promotions) {
 
     return promotionStrategies.stream()
